+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/JbodStorageTest.java
@@ -159,8 +159,6 @@ public class JbodStorageTest {
         });
         createAsync.await();
 
-        Set<String> expectedPvcs = expectedPvcs();
-
         // trying to add a new volume to the JBOD storage
         volumes.add(new PersistentClaimStorageBuilder()
                 .withId(2)
@@ -174,6 +172,8 @@ public class JbodStorageTest {
                     .endKafka()
                 .endSpec()
                 .build();
+
+        Set<String> expectedPvcs = expectedPvcs(changedKafka);
 
         Crds.kafkaOperation(mockClient).inNamespace(NAMESPACE).withName(NAME).patch(changedKafka);
 
@@ -200,8 +200,6 @@ public class JbodStorageTest {
         });
         createAsync.await();
 
-        Set<String> expectedPvcs = expectedPvcs();
-
         // trying to remove a volume from the JBOD storage
         volumes.remove(0);
 
@@ -212,6 +210,8 @@ public class JbodStorageTest {
                     .endKafka()
                 .endSpec()
                 .build();
+
+        Set<String> expectedPvcs = expectedPvcs(changedKafka);
 
         Crds.kafkaOperation(mockClient).inNamespace(NAMESPACE).withName(NAME).patch(changedKafka);
 
@@ -238,7 +238,7 @@ public class JbodStorageTest {
         });
         createAsync.await();
 
-        Set<String> expectedPvcs = expectedPvcs();
+        Set<String> expectedPvcs = expectedPvcs(this.kafka);
 
         // trying to update id for a volume from in the JBOD storage
         volumes.get(0).setId(3);
@@ -263,9 +263,9 @@ public class JbodStorageTest {
         });
     }
 
-    private Set<String> expectedPvcs() {
+    private Set<String> expectedPvcs(Kafka kafka) {
         Set<String> expectedPvcs = new HashSet<>();
-        for (int i = 0; i < this.kafka.getSpec().getKafka().getReplicas(); i++) {
+        for (int i = 0; i < kafka.getSpec().getKafka().getReplicas(); i++) {
             int podId = i;
             for (SingleVolumeStorage volume : this.volumes) {
                 if (volume instanceof PersistentClaimStorage) {