+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.12.0
 
+* Adding and removing volumes in Jbod Storage
 * Log at the warn level when a custom resource uses deprecated or unknown properties
 * Reduce the number of images needed to run Strimzi to just two: `kafka` and `operator`.
 * Add support for unprivileged users to install the operator with Helm